Config Structs are now passed by refrence and its all kosher
This was done by using some pointer and reference magic and then
changing the char * to QString

This should probably be worked around such that it can still be a char*
since now a conversion is neccesary to a tempvalue whenever making a
modbus variable.

Additionally it relys on Qt var in critical points but for now this is
acceptable.