Update to the Online Backpropover Modbus TCP
So tried to add a bit to the runmodbusOnline Method

So far we still need to look at exactly how to use efficiency to
actually make the backpropogation work.

Basically the attempt here was to pass outputs and inputs to an
overloaded backpropogation method and then calculate the inverse of the
total system efficiency. This would then be put in place of the error.

It seemed like this would be a good idea until the target values were
observed.. basically this process doesnt account for the gradient
calculation method as they use the targetval vector (pretty much
irrelevant here) and even if the efficiency was put here it still likely
wouldnt really work with the m_output private var.